Går fra Long til Int for sats i finnBehadlingerForSatsendring (#1854)
* Går fra Long til Int for sats i finnBehadlingerForSatsendring

+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/autovedtak/satsendring/SatsendringService.kt
@@ -34,7 +34,7 @@ class SatsendringService(
      * man sjekke om den inaktive behandlingen blir etterfulgt av revurdering som fjerner behovet.
      */
     fun finnBehandlingerForSatsendring(
-        gammelSats: Long,
+        gammelSats: Int,
         satsendringMåned: YearMonth
     ): List<Long> =
         behandlingRepository.finnBehadlingerForSatsendring(
